Fix cproto recipe
The main cproto download site only contains the latest version
and the name of the file is cproto.tar.gz. A few updates have
been released since the version we were using, hence the tarball
md5 and sha256 no longer match. In an attempt to modify these
for the latest release, the autotools build patch no longer applies.

In order to fix the issue, there is an archives site which has
copies of all previous releases, as well as the current release
so we switch to using the archive URI and select the archiver we
require.